refactor: PostCardStatus 경로를 Enums로 변경
