+++ src/main/java/com/bookbla/americano/domain/postcard/enums/PostcardStatus.java
@@ -1,4 +1,4 @@
-package com.bookbla.americano.domain.postcard.repository.entity;
+package com.bookbla.americano.domain.postcard.enums;
 
 public enum PostcardStatus {
     PENDING, ACCEPT, REFUSED, EXPIRED, ALL_WRONG
